pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hrl3d"
version = "0.3.0"
description = "HRL-driven 3D parallel planning (DP/TP/PP) for heterogeneous clusters; includes probing, grouping, GPT-2 introspection, simulated cost model, and YAML plan export."
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "numpy>=1.23",
]

[project.optional-dependencies]
probe = ["psutil>=5.9"]
gnn = ["torch-geometric>=2.3"]
rl = ["torch>=2.0"]
hf = ["transformers>=4.35", "huggingface_hub>=0.17", "safetensors>=0.4"]

[project.scripts]
hrl3d = "hrl3d.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
